Fix: Apply formatting fixes from CI
- Ran `cargo fmt` to fix import ordering and line wrapping in `src/semantic/model.rs`, `src/semantic/conversion.rs`, `src/semantic/expressions.rs` and `src/semantic/assembler.rs`.
- This ensures the codebase passes the formatting check.
