+++ src/parser/mod.rs
@@ -27,6 +27,7 @@ pub(crate) mod declarations;
 pub(crate) mod expressions;
 pub mod grammar;
 pub mod numerals;
+pub mod recursion;
 pub(crate) mod statements;
 
 use self::grammar::{Rule, parse as grammar_parse};
@@ -62,7 +63,7 @@ pub fn parse(source: &str) -> Result<Program, GlossaError> {
 /// Build an AST from source code
 fn parse_source(source: &str) -> Result<Program, ParseError> {
     // Check recursion depth before parsing to prevent stack overflow
-    common::check_recursion_depth(source)?;
+    recursion::check_recursion_depth(source)?;
 
     let pairs = grammar_parse(source).map_err(|e| ParseError::PestError(e.to_string()))?;
 
@@ -241,7 +242,7 @@ mod tests {
         // 500 nested parentheses (should pass check, though pest might fail to parse empty parens)
         let source = "(".repeat(500) + &")".repeat(500);
         // We only care about the recursion check here
-        let result = common::check_recursion_depth(&source);
+        let result = recursion::check_recursion_depth(&source);
         assert!(result.is_ok());
     }
 
@@ -249,7 +250,7 @@ mod tests {
     fn test_recursion_limit_ignored_in_string() {
         // Parentheses inside string literal shouldn't count
         let source = "«".to_string() + &"(".repeat(600) + "»";
-        let result = common::check_recursion_depth(&source);
+        let result = recursion::check_recursion_depth(&source);
         assert!(result.is_ok());
     }
 
@@ -257,7 +258,7 @@ mod tests {
     fn test_recursion_limit_ignored_in_comment() {
         // Parentheses inside comment shouldn't count
         let source = "// ".to_string() + &"(".repeat(600);
-        let result = common::check_recursion_depth(&source);
+        let result = recursion::check_recursion_depth(&source);
         assert!(result.is_ok());
     }
 
@@ -271,7 +272,7 @@ mod tests {
             + &"]".repeat(101)
             + &"}".repeat(200)
             + &")".repeat(200);
-        let result = common::check_recursion_depth(&source);
+        let result = recursion::check_recursion_depth(&source);
         assert!(matches!(
             result,
             Err(ParseError::RecursionLimitExceeded(500))
@@ -284,7 +285,7 @@ mod tests {
         // (((...))) then (((...))) - sequential, not nested
         let part = "(".repeat(400) + &")".repeat(400);
         let source = part.clone() + &part;
-        let result = common::check_recursion_depth(&source);
+        let result = recursion::check_recursion_depth(&source);
         assert!(result.is_ok());
     }
 }
